+++ scst/include/backport.h
@@ -142,7 +142,7 @@ static inline unsigned int cpumask_next(int n, const cpumask_t *srcp)
 	/* -1 is a legal arg here. */
 	if (n != -1)
 		cpumask_check(n);
-	return find_next_bit(cpumask_bits(srcp), nr_cpumask_bits, n+1);
+	return (int)find_next_bit(cpumask_bits(srcp), nr_cpumask_bits, n+1);
 }
 
 /**
